Changed from @Data to GettersAndSetters in entities because @Data generates toString and this makes a infinite recursion loop if there is bidirectional relationship. Also the error for infinite recursion is usually this org.hibernate.collection.spi.PersistentSet.toString(PersistentSet.java:298)\r\n\tat java.base/java.lang.String.valueOf(String.java:4220)\r\n\tat dev.sanderk.home_media_server.model.Movie.toString(Movie.java:8)\r\n\tat java.base/java.lang.String.valueOf(String.java:4220)\r\n\tat
But many rows of that same bs.
Also I read that Data is usually used for DTO's

Also tested out Slice, Page for Pageable. My idea is that I will start using Slice because it is straightforward and a client can see metadata that helps to show content more efficiently. I'm still testing it though, so I will implement it tomorrow